requirements.txt
nltk
cyhunspell
pycld3
nnsplit
